requirements.txt
GitPython==3.1.43
requests==2.32.3
